missed one

+++ pyro/mesh/array_indexer.py
@@ -502,7 +502,7 @@ class ArrayIndexerFC(ArrayIndexer):
         """
 
         if fmt is None:
-            if self.dtype == np.int64:
+            if self.dtype == int:
                 fmt = "%4d"
             elif self.dtype == np.float64:
                 fmt = "%10.5g"
